% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical_data}
\alias{clinical_data}
\title{Swiss clinical data}
\format{
A data frame with 4037 obs. of  6 variables.
\describe{
  \item{division}{factor with 27 levels, the Swiss canton}
  \item{date}{factor with 424 levels, representing the date of sampling}
  \item{HV6970del}{the count of sequences with the S-gene HV69-70 deletion}
  \item{ORF1adel}{the count of sequences with the ORF1a deletion}
  \item{ORF1adel}{the count of sequences determined to be B.1.1.7 (Alpha)}
  \item{ORF1adel}{the total count of sequences}
}
}
\source{
\url{cov-spectrum.ethz.ch}
}
\usage{
clinical_data
}
\description{
A dataset containing Swiss clinical sequencing data used in the publication:
Caduff, Lea, David Dreifuss, Tobias Schindler, Alexander J. Devaux,
Pravin Ganesanandamoorthy, Anina Kull, Elyse Stachler et al.
"Inferring Transmission Fitness Advantage of SARS-CoV-2 Variants of Concern
in Wastewater Using Digital PCR."
 medRxiv (2021).
}
\keyword{datasets}
