% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dPCRdat_HV6970}
\alias{dPCRdat_HV6970}
\title{Zurich HV69-70 dPCR data}
\format{
A data frame with 64 obs. of  6 variables.
\describe{
  \item{date}{Date, representing the date of sampling}
  \item{dayssince}{number of days since 2020-12-07}
  \item{Dilutionfactor}{factor used in dilution,
  useful for calculating viral RNA concentration}
  \item{x0}{double negative droplets count}
  \item{x1}{single positive droplets count}
  \item{x2}{double positive droplets count}
}
}
\source{
\doi{10.25678/0005ET}
}
\usage{
dPCRdat_HV6970
}
\description{
A dataset containing Zurich HV69-70 dPCR data used in the publication:
Caduff, Lea, David Dreifuss, Tobias Schindler, Alexander J. Devaux,
Pravin Ganesanandamoorthy, Anina Kull, Elyse Stachler et al.
"Inferring Transmission Fitness Advantage of SARS-CoV-2 Variants of Concern
in Wastewater Using Digital PCR."
 medRxiv (2021).
}
\keyword{datasets}
