% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pl.R
\name{fit_pl}
\alias{fit_pl}
\title{fit_pl}
\usage{
fit_pl(
  ydata,
  xdata,
  weights = 1,
  fitfun = pl3model,
  likelihood = loglik_binom_n,
  starting_values = c(a = 0.5, b = -10, c = 0.2),
  lower = c(-Inf, -Inf, 0),
  upper = c(Inf, Inf, 1)
)
}
\arguments{
\item{ydata}{response variable}

\item{xdata}{independent variable}

\item{weights}{weights vector (optional).}

\item{fitfun}{fitting function}

\item{likelihood}{likelihood function}

\item{starting_values}{(named) vector of starting values for optimization}

\item{lower}{vector of lower bounds for parameters}

\item{upper}{vector of upper bounds for parameters}
}
\value{
pl_fit object
}
\description{
fit_pl
}
