% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pl.R
\name{logist_confint}
\alias{logist_confint}
\title{logist_confint}
\usage{
logist_confint(x)
}
\arguments{
\item{x}{an object of class \code{pl_fit}}
}
\value{
list of confidence intervals
}
\description{
logist_confint
}
