% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pl.R
\name{logist_predint}
\alias{logist_predint}
\title{logist_predint}
\usage{
logist_predint(x, ...)
}
\arguments{
\item{x}{an object of class \code{pl_fit}}

\item{...}{additional argument(s) for methods}
}
\value{
list of prediction intervals
}
\description{
logist_predint
}
