% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lower_e}
\alias{lower_e}
\title{lower_e}
\usage{
lower_e(x, level = 0.95)
}
\arguments{
\item{x}{vector or matrix of droplet counts}

\item{level}{1-alpha confidence level}
}
\value{
lower confidence bounds
}
\description{
lower_e
}
