% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lrt_confint}
\alias{lrt_confint}
\title{lrt_confint}
\usage{
lrt_confint(x, level = 0.95)
}
\arguments{
\item{x}{vector or matrix of droplet counts}

\item{level}{1-alpha confidence level}
}
\value{
data.frame of lower and upper confidence bounds
}
\description{
lrt_confint
}
