% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfuns.R
\name{pl2model}
\alias{pl2model}
\title{pl2model}
\usage{
pl2model(x, a, b)
}
\arguments{
\item{x}{time}

\item{a}{rate parameter}

\item{b}{bias parameter}
}
\value{
a two parametric logistic curve
}
\description{
pl2model
}
\details{
returns a logistic curve with bottom asymptote at 0, such that: \eqn{r(x;a,b) = \frac{\exp(ax+b)}{1+exp(a*x+b)}}
}
