% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfuns.R
\name{pl3model}
\alias{pl3model}
\title{pl3model}
\usage{
pl3model(x, a, b, c)
}
\arguments{
\item{x}{time}

\item{a}{rate parameter}

\item{b}{bias parameter}

\item{c}{background parameter}
}
\value{
a three parametric logistic curve
}
\description{
pl3model
}
\details{
returns a logistic curve with bottom asymptote not at 0 but at parameter c, such that: \eqn{r(x;a,b,c) = c + (1-c) \frac{\exp(ax+b)}{1+exp(a*x+b)}}
}
