% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{upper_e}
\alias{upper_e}
\title{upper_e}
\usage{
upper_e(x, level = 0.95)
}
\arguments{
\item{x}{vector or matrix of droplet counts}

\item{level}{1-alpha confidence level}
}
\value{
upper confidence bounds
}
\description{
upper_e
}
