% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pl.R
\name{fit_pl}
\alias{fit_pl}
\alias{coef.pl_fit}
\alias{se.pl_fit}
\alias{fitted.pl_fit}
\alias{confint.pl_fit}
\alias{predict.pl_fit}
\alias{logist_confint.pl_fit}
\alias{logist_predint.pl_fit}
\title{fit_pl}
\usage{
fit_pl(
  ydata,
  xdata,
  weights = 1,
  fitfun = pl3model,
  likelihood = loglik_binom_n,
  quasi = F,
  starting_values = c(a = 0.5, b = -10, c = 0.2),
  lower = c(-Inf, -Inf, 0),
  upper = c(Inf, Inf, 1)
)

\method{coef}{pl_fit}(object, ...)

\method{se}{pl_fit}(x)

\method{fitted}{pl_fit}(object, ...)

\method{confint}{pl_fit}(object, parm, level = 0.95, ...)

\method{predict}{pl_fit}(object, newdata = NULL, scale = c("linear", "logit"), se = TRUE, ...)

\method{logist_confint}{pl_fit}(x, newdata = NULL, level = 0.95, ...)

\method{logist_predint}{pl_fit}(
  x,
  newdata = NULL,
  level = 0.95,
  likelihood = loglik_binom_n,
  ...
)
}
\arguments{
\item{ydata}{response variable}

\item{xdata}{independent variable}

\item{weights}{weights vector (optional).}

\item{fitfun}{fitting function}

\item{likelihood}{likelihood function}

\item{quasi}{compute overdispersion and use quasilikelihood model to build confidence intervals}

\item{starting_values}{(named) vector of starting values for optimization}

\item{lower}{vector of lower bounds for parameters}

\item{upper}{vector of upper bounds for parameters}

\item{object}{an object of class \code{pl_fit}}

\item{...}{additional argument(s)}

\item{x}{an object of class \code{pl_fit}}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{confidence level 1-alpha for prediction}

\item{newdata}{an optional vector of observed independent variables with which to predict}

\item{scale}{scale on which to predict, either linear or logit scale}

\item{se}{whether to return the standard errors of the predictions or not}
}
\value{
pl_fit object
}
\description{
fit_pl
}
