% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfuns.R
\name{loglik_binom_n}
\alias{loglik_binom_n}
\title{loglik_binom_n}
\usage{
loglik_binom_n(y, p)
}
\arguments{
\item{y}{empirical success frequency}

\item{p}{probability of success}
}
\value{
log-likelihood kernel
}
\description{
loglik_binom_n
}
