% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods (1).R, R/likelihoods.R
\name{loglik_trinom_prof}
\alias{loglik_trinom_prof}
\title{loglik_trinom_prof}
\usage{
loglik_trinom_prof(y, r)

loglik_trinom_prof(y, r)
}
\arguments{
\item{y}{data matrix of counts}

\item{r}{expected ratio of concentrations}
}
\value{
log-likelihood kernel

log-likelihood kernel
}
\description{
loglik_trinom_prof

loglik_trinom_prof
}
